pyproject.toml
[project]
name = "album-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imagehash>=4.3.2",
    "networkx>=3.5",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "piexif>=1.1.3",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "streamlit>=1.50.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
